refactor: move logic for game modes to separate files
